% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{causal_lp-defunct}
\alias{causal_lp-defunct}
\title{Get the log probability of each element of a vector of words (or phrases) using a causal transformer}
\description{
This function is defunct. Use \code{causal_targets_pred()} or \code{causal_words_pred()} instead.
}
\seealso{
\code{\link{pangoling-defunct}}
}
\keyword{internal}
