% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{causal_lp_mats-defunct}
\alias{causal_lp_mats-defunct}
\title{Get a list of matrices with the log probabilities of possible words given their previous context using a causal transformer}
\description{
This function is defunct. Use \code{causal_pred_mats()} instead.
}
\seealso{
\code{\link{pangoling-defunct}}
}
\keyword{internal}
