% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{causal_lp_mats-deprecated}
\alias{causal_lp_mats-deprecated}
\title{Get a list of matrices with the log probabilities of possible words
given their previous context using a causal transformer}
\description{
This function is deprecated. Use \code{causal_pred_mats()} instead.
}
\seealso{
\code{\link{pangoling-deprecated}}
}
\keyword{internal}
