% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{causal_next_tokens_tbl-deprecated}
\alias{causal_next_tokens_tbl-deprecated}
\title{Get the possible next tokens and their log probabilities for its previous context}
\usage{
causal_next_tokens_tbl(context, model, checkpoint = NULL, add_special_tokens = NULL, config_model = NULL, config_tokenizer = NULL)
}
\description{
This function is deprecated. Use \code{causal_next_tokens_pred_tbl()} instead.
}
\seealso{
\code{\link{pangoling-deprecated}}
}
\keyword{internal}
