% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_causal.R
\name{causal_next_tokens_tbl}
\alias{causal_next_tokens_tbl}
\title{Get the possible next tokens and their log probabilities its previous context using a causal transformer}
\usage{
causal_next_tokens_tbl(
  context,
  model = getOption("pangoling.causal.default"),
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL
)
}
\arguments{
\item{context}{The context.}

\item{model}{Name of a pretrained model stored locally on the (huggingface.co).}

\item{add_special_tokens}{Whether to include beginning of text special tokens. By default  acts as the \href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoTokenizer}{AutoTokenizer}.}

\item{config_model}{List with other arguments that control how the model from hugging face is accessed.}

\item{config_tokenizer}{List with other arguments that control how the tokenizer from hugging face is accessed.}
}
\value{
A table with possible next tokens and their log-probabilities.
}
\description{
Get the possible next tokens and their log probabilities based on its previous context using a causal transformer model from \href{https://huggingface.co}{Hugging Face}.
}
\details{
For more about causal models, see \href{https://huggingface.co/course/chapter7/6}{chapter 7 of hugging face documentation}.

If not specified, the causal model that will be used is the one set in specified in the global option \code{pangoling.causal.default}, this can be accessed via \code{getOption("pangoling.causal.default")} (by default "gpt2"). To change the default option use \code{options(pangoling.causal.default = "newcausalmodel")}.

A list of possible causal models can be found in \href{https://huggingface.co/}{hugging face website}.

Using the  \code{config_model} and \code{config_tokenizer} arguments, it's possible to control how the model and tokenizer from hugging face is accessed, see the python method \href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoProcessor.from_pretrained}{\code{from_pretrained}} for details. In case of errors check the status of \url{https://status.huggingface.co/}
}
\section{More examples}{

See the  \href{https://bruno.nicenboim.me/pangoling/articles/intro-gpt2.html}{online article} in pangoling website for more examples.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
causal_next_tokens_tbl(
  context = "The apple doesn't fall far from the",
  model = "gpt2"
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other causal model functions: 
\code{\link{causal_config}()},
\code{\link{causal_lp_mats}()},
\code{\link{causal_lp}()},
\code{\link{causal_preload}()},
\code{\link{causal_tokens_lp_tbl}()}
}
\concept{causal model functions}
