% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{causal_tokens_lp_tbl-defunct}
\alias{causal_tokens_lp_tbl-defunct}
\title{Get the log probability of each token in a sentence (or group of sentences) using a causal transformer}
\description{
This function is defunct. Use \code{causal_tokens_pred_lst()} instead.
}
\seealso{
\code{\link{pangoling-defunct}}
}
\keyword{internal}
