% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{causal_tokens_lp_tbl}
\alias{causal_tokens_lp_tbl}
\title{Get the log probability of each token in a sentence (or group of sentences) using a causal transformer}
\usage{
causal_tokens_lp_tbl(
  texts,
  model = getOption("pangoling.causal.default"),
  checkpoint = NULL,
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL,
  batch_size = 1,
  .id = NULL
)
}
\arguments{
\item{texts}{Vector or list of texts.}

\item{model}{Name of a pre-trained model or folder.}

\item{checkpoint}{Folder of a checkpoint.}

\item{add_special_tokens}{Whether to include special tokens. It has the
same default as the
\href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoTokenizer}{AutoTokenizer} method in Python.}

\item{config_model}{List with other arguments that control how the
model from Hugging Face is accessed.}

\item{config_tokenizer}{List with other arguments that control how the tokenizer from Hugging Face is accessed.}

\item{batch_size}{Maximum size of the batch. Larges batches speedup
processing but take more memory.}

\item{.id}{Name of the column with the sentence id.}
}
\value{
A table with token names (\code{token}), log-probability (\code{lp}) and optionally sentence id.
}
\description{
Get the log probability of each token in a sentence (or group of sentences) using a causal transformer model.
}
\details{
A causal language model (also called GPT-like, auto-regressive, or decoder
model) is a type of large language model usually used for text-generation
that can predict the next word (or more accurately in fact token) based
on a preceding context.

If not specified, the causal model that will be used is the one set in
specified in the global option \code{pangoling.causal.default}, this can be
accessed via \code{getOption("pangoling.causal.default")} (by default
"gpt2"). To change the default option
use \code{options(pangoling.causal.default = "newcausalmodel")}.

A list of possible causal models can be found in
\href{https://huggingface.co/models?pipeline_tag=text-generation}{Hugging Face website}.

Using the  \code{config_model} and \code{config_tokenizer} arguments, it's possible to
control how the model and tokenizer from Hugging Face is accessed, see the
Python method
\href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoProcessor.from_pretrained}{\code{from_pretrained}}
for details.

In case of errors when a new model is run, check the status of
\url{https://status.huggingface.co/}
}
\section{More examples}{

See the
\href{https://bruno.nicenboim.me/pangoling/articles/intro-gpt2.html}{online article}
in pangoling website for more examples.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
causal_tokens_lp_tbl(
  texts = c("The apple doesn't fall far from the tree."),
  model = "gpt2"
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other causal model functions: 
\code{\link{causal_config}()},
\code{\link{causal_lp}()},
\code{\link{causal_lp_mats}()},
\code{\link{causal_next_tokens_pred_tbl}()},
\code{\link{causal_next_tokens_tbl}()},
\code{\link{causal_pred_mats}()},
\code{\link{causal_preload}()},
\code{\link{causal_targets_pred}()},
\code{\link{causal_tokens_pred_tbl}()},
\code{\link{causal_words_pred}()}
}
\concept{causal model functions}
