% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_frank2013_et_fix}
\alias{data_frank2013_et_fix}
\title{ET data  of Frank et al ...}
\format{
A data frame with  rows and  variables:
\itemize{
\item \code{subj}: Subject
\item \code{sent_id}: Sentence unique id.
\item \code{gaze_x}: Horizontal pixel coordinate of fixation location (0 = left edge; 39 = left margin; 1024 = right edge) Each letter (including space and punctuation) was 14 pixels wide.
\item \code{gaze_y}: Vertical pixal coordinate of fixation location (0 = top of display; 768 = bottom of display)
\item \code{fix_duration}: Duration of fixation in msec
\item \code{letter_n}: Fixated letter position in sentence (including spaces), even if the fixation was to the left of the left margin or to the right of the last letter. Position 0 is directly to the left of the first letter.
\item \code{word_n}: Position of the word in the sentence
\item \code{word}: Presented word
\item \code{blink}: Indicates whether a blink was detected directly 'before' or 'after' the current fixation (or 'both' before and after).
}

NOTE -- Within each sentence, fixations are listed in the order in which they occurred. If any fixation during sentence presentation was registered as being outside the display bounds, this was regarded as a tracking error and no data is presented for that sentence.
NOTE -- The letter and word position are NaN if gaze_y is outside the range (300,475) because that was considered too far above or below the presented sentence.
}
\usage{
data_frank2013_et_fix
}
\description{
ET data  of Frank et al ...
}
\keyword{datasets}
