% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_frank2013_spr}
\alias{data_frank2013_spr}
\title{SPR  of Frank et al ...}
\format{
An object of class \code{tidytable} (inherits from \code{data.table}, \code{data.frame}) with 165040 rows and 12 columns.
}
\usage{
data_frank2013_spr
}
\description{
Cleaned version of dataset \link{data_frank2013_spr_complete}: No sentences with typos, only native speakers, and subjects with accuracy over 0.8.
}
\keyword{datasets}
