% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_frank2013_spr_complete}
\alias{data_frank2013_spr_complete}
\title{Self-paced reading data of Frank, Fernandez Monsalve, Thompson et al. (2013)}
\format{
A data frame with  rows and  variables:
\itemize{
\item \code{subj}: Subject id.
\item \code{sent_id}: Sentence unique id.
\item \code{sent_n}: Position of sentence in the presentation sequence.
\item \code{answer_time}: Time in ms between question presentation and response, or NaN if there was no question.
\item \code{word_n}: Position of the word in the sentence.
\item \code{word}: Presented word.
\item \code{RT}: Time in ms between word presentation and key press.
\item \code{acc_comprehension}: Correctness of response to comprehension question.
\item \code{age}: Subject's age in years.
\item \code{age_en}: Age at which subject began learning English (0 for native speakers).
\item \code{sex}: Subject's sex (f/m).
\item \code{hand}: Subject's handedness (r/l).
\item \code{correct_perc}: Fraction of correct responses to comprehension questions.
\item \code{typo}: Whether there is a typo in the word.
}
}
\usage{
data_frank2013_spr_complete
}
\description{
A dataset containing the self-paced reading data of Frank, Fernandez Monsalve, Thompson et al. (2013). Notice that after collecting the self-paced-reading data, three typos were found:
\itemize{
\item sentence 43  : "Scott" was "Sott"
\item sentence 269 : "at" was "that"
\item sentence 337 : "Margaret" was "Margeret"
}
}
\references{
\link{1}
S. L. Frank, I. Fernandez Monsalve, R. L. Thompson, et al.
``Reading Time Data for Evaluating Broad-Coverage Models of English Sentence Processing''.
In: \emph{Behavior Research Methods} 45.4 (dec. 01, 2013), pp. 1182--1190.
ISSN: 1554-3528.
DOI: 10.3758/s13428-012-0313-y.
}
\seealso{
\link{data_frank2013_stimuli} for the enriched stimuli,  \link{data_frank2013_spr} for a cleaned version of this dataset,  \link{data_frank2013_et_fix} for the eye-tracking version of this dataset with the fixation information, and \link{data_frank2013_et} for a preprocessed eye-tracking dataset.

Other datasets: 
\code{\link{data_frank2013_et_rt}},
\code{\link{data_frank2013_spr}},
\code{\link{data_frank2013_stimuli}},
\code{\link{data_provo_cloze}}
}
\concept{datasets}
\keyword{datasets}
