% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_provo_cloze}
\alias{data_provo_cloze}
\title{Cloze probabilities from the Provo Corpus.}
\format{
A data frame with 2730 rows and 16 variables:
\itemize{
\item \code{word_unique_id}: DESCRIPTION
\item \code{text_id}: DESCRIPTION
\item \code{word_n}: DESCRIPTION
\item \code{word}: DESCRIPTION
\item \code{word_cleaned}: DESCRIPTION
\item \code{word_length}: DESCRIPTION
\item \code{total_response_count}: DESCRIPTION
\item \code{cp}: DESCRIPTION
\item \code{modal_response}: DESCRIPTION
\item \code{pos_claws}: DESCRIPTION
\item \code{word_content_or_function}: DESCRIPTION
\item \code{word_pos}: DESCRIPTION
\item \code{pos_cp}: DESCRIPTION
\item \code{lsa_context_score}: DESCRIPTION
\item \code{lsa_response_match_score}: DESCRIPTION
\item \code{text}: DESCRIPTION
}
}
\source{
\url{https://osf.io/sjefs//}
}
\usage{
data_provo_cloze
}
\description{
A dataset containing the .. reference
}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> No results.
}\if{html}{\out{</div>}}

list()
}
\seealso{
Other datasets: 
\code{\link{data_frank2013_et_rt}},
\code{\link{data_frank2013_spr_complete}},
\code{\link{data_frank2013_spr}},
\code{\link{data_frank2013_stimuli}}
}
\concept{datasets}
\keyword{datasets}
