% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtlex.R
\name{freq_sources}
\alias{freq_sources}
\title{List of frequency sources}
\usage{
freq_sources()
}
\value{
list of frequency sources for \code{\link[=get_all_word_freq_tbl]{get_all_word_freq_tbl()}} and \code{\link[=get_word_freq_tbl]{get_word_freq_tbl()}}
}
\description{
List of frequency sources
}
\seealso{
Other frequency: 
\code{\link{get_word_freq_tbl}()}
}
\concept{frequency}
