% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtlex.R
\name{get_all_word_freq_tbl}
\alias{get_all_word_freq_tbl}
\title{Title}
\usage{
get_all_word_freq_tbl(
  x,
  source = "subtlex",
  language = "english",
  location = "US"
)
}
\arguments{
\item{location}{}
}
\value{

}
\description{
Title
}
