% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_causal.R
\name{get_causal_entropy}
\alias{get_causal_entropy}
\title{Title}
\usage{
get_causal_entropy(x, by = rep(1, length(x)), model = "gpt2", eot = 0)
}
\arguments{
\item{model}{}
}
\description{
Incorrect for multi-token words.
}
