% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_causal.R
\name{get_causal_model_config}
\alias{get_causal_model_config}
\title{Returns the configuration of a causal model}
\usage{
get_causal_model_config(model = "gpt2", config = NULL)
}
\arguments{
\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}
}
\value{
A list with the configuration of the model
}
\description{
Returns the configuration of a causal model
}
