% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_causal.R
\name{get_causal_next_tokens_tbl}
\alias{get_causal_next_tokens_tbl}
\title{Get the possible next tokens and their log probabilities its previous context using a causal transformer.}
\usage{
get_causal_next_tokens_tbl(
  context,
  model = "gpt2",
  add_bos_token = NULL,
  config_model = list(),
  config_tokenizer = list()
)
}
\arguments{
\item{context}{Context}

\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}

\item{add_bos_token}{Whether to include beginning of text special tokens. See details}

\item{config_model}{list with other arguments that control how the model from hugging face is accessed.}

\item{config_tokenizer}{list with arguments that control how the model from hugging face is accessed.}
}
\value{
a table with possible next tokes and their log-probabilities
}
\description{
Get the possible next tokens and their log probabilities its previous context using a causal transformer model from huggingface.com.
}
\details{
\code{add_bos_token} by default  acts as the \href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoTokenizer}{AutoTokenizer}. Using \code{...} it's possible to control how the model from hugging face is accessed, see \href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoProcessor.from_pretrained}{from_pretrained} for details.
}
