% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_utils.R
\name{get_id}
\alias{get_id}
\title{Get ids (without adding special characters at beginning or end?)}
\usage{
get_id(x, model = "gpt2", add_bos_token = NULL, config = NULL)
}
\description{
Get ids (without adding special characters at beginning or end?)
}
