% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_masked.R
\name{get_masked_log_prob_mat}
\alias{get_masked_log_prob_mat}
\title{Title}
\usage{
get_masked_log_prob_mat(
  x,
  by = rep(1, length(x)),
  model = "distilbert-base-uncased",
  max_batch_size = 50,
  window_stride = 5,
  n_plus = 3,
  max_tokens = max_tokens_masked(model)
)
}
\arguments{
\item{model}{}
}
\description{
Title
}
