% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_masked.R
\name{get_masked_lp}
\alias{get_masked_lp}
\title{Get the log probability of each word phrase of a vector given its previous context using a transformer model from huggingface.co.}
\usage{
get_masked_lp(
  x,
  .by = rep(1, length(x)),
  ignore_regex = "",
  model = "bert-base-uncased",
  add_special_tokens = NULL,
  stride = 1,
  config_model = NULL,
  config_tokenizer = NULL
)
}
\arguments{
\item{x}{Vector of words, phrases or texts.}

\item{.by}{Vector that indicates how the text should be split.}

\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}
}
\value{
a vector of log probabilities.
}
\description{
In case of errors check the status of https://status.huggingface.co/
}
