% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_masked.R
\name{get_masked_tokens_tbl}
\alias{get_masked_tokens_tbl}
\title{get_masked_tokens_tbl}
\usage{
get_masked_tokens_tbl(
  masked_sentences,
  model = "bert-base-uncased",
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL
)
}
\arguments{
\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}

\item{context}{Context}
}
\value{
A table
}
\description{
get_masked_tokens_tbl
}
