% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{get_regex_fpmw_lst}
\alias{get_regex_fpmw_lst}
\title{Get the word frequency per million words using regex}
\usage{
get_regex_fpmw_lst(x, source = "SUBTLEX_US", ignore_case = TRUE)
}
\arguments{
\item{remove_punc}{}
}
\value{

}
\description{
Get the word frequency per million words using regex
}
\examples{
 get_regex_fpmw_lst(x = c("car", "cat", "bubble"))

}
