% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{get_tr_log_prob}
\alias{get_tr_log_prob}
\title{Get the log probability of each word phrase of a vector given its previous context using a transformer model from huggingface.co/.}
\usage{
get_tr_log_prob(x, by = rep(1, length(x)), eot = 0, model = "gpt2")
}
\arguments{
\item{x}{Vector of words, phrases or texts.}

\item{by}{Vector that indicates how the text should be split.}

\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}
}
\value{
a vector of log probabilities.
}
\description{
In case of errors check the status of https://status.huggingface.co/
}
