% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{get_tr_log_prob_mat}
\alias{get_tr_log_prob_mat}
\title{Title}
\usage{
get_tr_log_prob_mat(x, by = rep(1, length(x)), eot = 0, model = "gpt2")
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Title
}
