% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{get_tr_next_tokens_tbl}
\alias{get_tr_next_tokens_tbl}
\title{Get the log probability of each word phrase of a vector given its previous context using a transformer model from huggingface.co/.}
\usage{
get_tr_next_tokens_tbl(context, model = "gpt2")
}
\arguments{
\item{context}{Context}

\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}
}
\value{

}
\description{
Get the log probability of each word phrase of a vector given its previous context using a transformer model from huggingface.co/.
}
