% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{get_word_feature}
\alias{get_word_feature}
\title{Get the word feature}
\usage{
get_word_feature(
  x,
  source = "SUBTLEX_US",
  ignore_case = TRUE,
  feature = "Zipf",
  remove_punc = "[[:punct:][:blank:]]"
)
}
\arguments{
\item{remove_punc}{}
}
\value{

}
\description{
Get the word feature
}
\examples{
 get_word_feature(x = c("car", "cat", "bubble"))
}
