% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtlex.R
\name{get_word_freq_tbl}
\alias{get_word_freq_tbl}
\title{Get a table with word frequencies.}
\usage{
get_word_freq_tbl(
  x,
  source = "subtlex",
  language = "english",
  location = "US",
  ignore_case = TRUE,
  remove_punc = "[[:punct:][:blank:]]",
  regex = FALSE
)
}
\arguments{
\item{regex}{From https://www.ugent.be/pp/experimentele-psychologie/en/research/documents/subtlexus
The word. This starts with a capital when the word more often starts with an uppercase letter than with a lowercase letter.
FREQcount. This is the number of times the word appears in the corpus (i.e., on the total of 51 million words).
CDcount. This is the number of films in which the word appears (i.e., it has a maximum value of 8,388).
FREQlow. This is the number of times the word appears in the corpus starting with a lowercase letter. This allows users to further match their stimuli.
CDlow. This is the number of films in which the word appears starting with a lowercase letter.
SUBTLWF. This is the word frequency per million words. It is the measure you would preferably use in your manuscripts, because it is a standard measure of word frequency independent of the corpus size. It is given with two digits precision, in order not to lose precision of the frequency counts.
Lg10WF. This value is based on log10(FREQcount+1) and has four digit precision. Because FREQcount is based on 51 million words, the following conversions apply for SUBTLEXUS:
  Lg10WF     SUBTLWF
1.00     0.2
2.00     2
3.00     20
4.00     200
5.00     2000
SUBTLCD indicates in how many percent of the films the word appears. This value has two-digit precision in order not to lose information.
Lg10CD. This value is based on log10(CDcount+1) and has four digit precision. It is the best value to use if you want to match words on word frequency. As CDcount is based on 8388 films, the following conversions apply:
  Lg10CD     SUBTLCD
0.95     0.1
1.93     1
2.92     10
3.92     100}
}
\value{

}
\description{
Get a table with word frequencies.
}
