% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minicons.R
\name{get_word_logprob}
\alias{get_word_logprob}
\title{Get the log probability of each word of a vector of words given its previous context.}
\usage{
get_word_logprob(
  x,
  by = rep(1, length(x)),
  model = "gpt2",
  device = "cpu",
  return_token = TRUE
)
}
\arguments{
\item{x}{Vector of words or texts.}

\item{by}{Vector that indicates how the text should be split}

\item{model}{Name of the model, should either be a path  to a model (.pt or .bin file) stored locally, or a  pretrained model stored on the Huggingface Model Hub.}

\item{device}{device type that the model should be loaded on, options: "cpu" or "cuda:{0, 1, ...}"?}
}
\value{

}
\description{
Get the log probability of each word of a vector of words given its previous context.
}
