% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{#' @noRd
message_debug <- function(...) {
if (options()$pangoling.verbose > 1) message(...)
}}
\usage{
x \%||\% y
}
\description{
#' @noRd
message_debug <- function(...) {
if (options()$pangoling.verbose > 1) message(...)
}
}
