% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangolang-package.R
\docType{package}
\name{pangolang-package}
\alias{pangolang}
\alias{pangolang-package}
\title{pangolang: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
