% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{pangoling-defunct}
\alias{pangoling-defunct}
\alias{masked_tokens_tbl}
\alias{masked_lp}
\alias{causal_next_tokens_tbl}
\alias{causal_lp}
\alias{causal_tokens_lp_tbl}
\alias{causal_lp_mats}
\title{Defunct functions in package \pkg{pangoling}.}
\usage{
masked_tokens_tbl(...)

masked_lp(...)

causal_next_tokens_tbl(...)

causal_lp(...)

causal_tokens_lp_tbl(...)

causal_lp_mats(...)
}
\description{
The functions listed below are now defunct.
When possible, alternative functions with similar functionality are also mentioned.
}
\section{\code{masked_tokens_tbl}}{

For \code{masked_tokens_tbl}, use \code{\link{masked_tokens_pred_tbl}}.
}

\section{\code{masked_lp}}{

For \code{masked_lp}, use \code{\link{masked_targets_pred}}.
}

\section{\code{causal_next_tokens_tbl}}{

For \code{causal_next_tokens_tbl}, use \code{\link{causal_next_tokens_pred_tbl}}.
}

\section{\code{causal_lp}}{

For \code{causal_lp}, use \code{\link{causal_targets_pred}} or \code{\link{causal_words_pred}}.
}

\section{\code{causal_tokens_lp_tbl}}{

For \code{causal_tokens_lp_tbl}, use \code{\link{causal_tokens_pred_lst}}.
}

\section{\code{causal_lp_mats}}{

For \code{causal_lp_mats}, use \code{\link{causal_pred_mats}}.
}

\keyword{internal}
