% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{pangoling-deprecated}
\alias{pangoling-deprecated}
\alias{masked_tokens_tbl}
\alias{masked_lp}
\alias{causal_next_tokens_tbl}
\alias{causal_lp}
\alias{causal_tokens_lp_tbl}
\alias{causal_lp_mats}
\title{Deprecated functions in package \pkg{pangoling}.}
\usage{
masked_tokens_tbl(
  masked_sentences,
  model = getOption("pangoling.masked.default"),
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL
)

masked_lp(
  l_contexts,
  targets,
  r_contexts,
  ignore_regex = "",
  model = getOption("pangoling.masked.default"),
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL
)

causal_next_tokens_tbl(
  context,
  model = getOption("pangoling.causal.default"),
  checkpoint = NULL,
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL
)

causal_lp(
  x,
  by = rep(1, length(x)),
  l_contexts = NULL,
  ignore_regex = "",
  model = getOption("pangoling.causal.default"),
  checkpoint = NULL,
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL,
  batch_size = 1,
  ...
)

causal_tokens_lp_tbl(
  texts,
  model = getOption("pangoling.causal.default"),
  checkpoint = NULL,
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL,
  batch_size = 1,
  .id = NULL
)

causal_lp_mats(
  x,
  by = rep(1, length(x)),
  sorted = FALSE,
  model = getOption("pangoling.causal.default"),
  checkpoint = NULL,
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL,
  batch_size = 1,
  ...
)
}
\description{
The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at \code{help("<function>-deprecated")}.
}
\section{\code{masked_tokens_tbl}}{

For \code{masked_tokens_tbl}, use \code{\link{masked_tokens_pred_tbl}}.
}

\section{\code{masked_lp}}{

For \code{masked_lp}, use \code{\link{masked_targets_pred}}.
}

\section{\code{causal_next_tokens_tbl}}{

For \code{causal_next_tokens_tbl}, use \code{\link{causal_next_tokens_pred_tbl}}.
}

\section{\code{causal_lp}}{

For \code{causal_lp}, use \code{\link{causal_targets_pred}} or \code{\link{causal_words_pred}}.
}

\section{\code{causal_tokens_lp_tbl}}{

For \code{causal_tokens_lp_tbl}, use \code{\link{causal_tokens_pred_tbl}}.
}

\section{\code{causal_lp_mats}}{

For \code{causal_lp_mats}, use \code{\link{causal_pred_mats}}.
}

\keyword{internal}
