% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangoling-package.R
\docType{package}
\name{pangoling-package}
\alias{pangoling}
\alias{pangoling-package}
\title{pangoling: Access to language model predictions in R}
\description{
Access to large language model word predictability in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://bruno.nicenboim.me/pangoling/}
}

}
\author{
\strong{Maintainer}: Bruno Nicenboim \email{bruno.nicenboim@gmail.com} (\href{https://orcid.org/0000-0002-5176-3943}{ORCID})

Other contributors:
\itemize{
  \item Chris Emmerly [contributor]
}

}
\keyword{internal}
