% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_causal.R
\name{perplexity}
\alias{perplexity}
\title{Calculates perplexity}
\usage{
perplexity(x, na.rm = FALSE, log.p = TRUE)
}
\arguments{
\item{x}{A vector of log-probabilities.}

\item{na.rm}{Should missing values (including NaN) be removed?}

\item{log.p}{If TRUE (default),  x are assumed to be log-transformed probabilities with base e, if FALSE x are assumed to be raw probabilities, alternatively log.p can be the base of other logarithmic transformations.}
}
\value{
The perplexity
}
\description{
Calculates perplexity of a vector of (log-)probabilities.
}
\details{

}
\examples{
probs <- c(.3,.5,.6)
perplexity(probs, log.p = FALSE)
lprobs <- log(probs)
perplexity(lprobs, log.p = TRUE)
}
