% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_utils.R
\name{var_to_py}
\alias{var_to_py}
\title{Sends a var to python
https://stackoverflow.com/questions/67562889/interoperability-between-python-and-r}
\usage{
var_to_py(var_name, x)
}
\description{
Sends a var to python
https://stackoverflow.com/questions/67562889/interoperability-between-python-and-r
}
