% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{versions}
\alias{versions}
\title{Show the version of python and relevant packages}
\usage{
versions()
}
\value{
a list with the python version, and the packages needed for \code{pangoling}
}
\description{
Show the version of python and relevant packages
}
\examples{

versions()

}
