% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_causal.R
\name{causal_config}
\alias{causal_config}
\title{Returns the configuration of a causal model}
\usage{
causal_config(model = "gpt2", config_model = NULL)
}
\arguments{
\item{model}{Name of a pretrained model stored locally on the (huggingface.co).}

\item{config_model}{List with other arguments that control how the model from hugging face is accessed.}
}
\value{
A list with the configuration of the model.
}
\description{
Returns the configuration of a causal model
}
\details{
For more about causal models, see \href{https://huggingface.co/course/chapter7/6}{chapter 7 of hugging face documentation}.  Using the  \code{config_model} and \code{config_tokenizer} arguments, it's possible to control how the model and tokenizer from hugging face is accessed, see the python method \href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoProcessor.from_pretrained}{\code{from_pretrained}} for details. In case of errors check the status of \url{https://status.huggingface.co/}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
causal_config(model = "gpt2")
\dontshow{\}) # examplesIf}
}
\seealso{
Other causal model functions: 
\code{\link{causal_lp_mats}()},
\code{\link{causal_lp}()},
\code{\link{causal_next_tokens_tbl}()},
\code{\link{causal_preload}()},
\code{\link{causal_tokens_lp_tbl}()}
}
\concept{causal model functions}
