% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{causal_lp-deprecated}
\alias{causal_lp-deprecated}
\title{Get the log probability of each element of a vector of words (or
phrases) using a causal transformer}
\description{
This function is deprecated. Use \code{causal_targets_pred()}
(supports \code{l_context} argument) or \code{causal_words_pred()} (supports \code{x} and
\code{by} arguments) instead.
}
\seealso{
\code{\link{pangoling-deprecated}}
}
\keyword{internal}
