% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{causal_next_tokens_tbl-defunct}
\alias{causal_next_tokens_tbl-defunct}
\title{Get the possible next tokens and their log probabilities for its previous context}
\value{
This function is defunct and returns an error.
}
\description{
This function is defunct. Use \code{causal_next_tokens_pred_tbl()} instead.
}
\seealso{
\code{\link{pangoling-defunct}}
}
\keyword{internal}
