% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_frank2013_et_fix}
\alias{data_frank2013_et_fix}
\title{Fixations positions of the eye-tracking data of \insertCite{Frank2013;textual}{pangoling}}
\format{
A data frame:
\itemize{
\item \code{subj}: Subject
\item \code{sent_id}: Sentence unique id.
\item \code{gaze_x}: Horizontal pixel coordinate of fixation location (0 = left edge; 39 = left margin; 1024 = right edge) Each letter (including space and punctuation) was 14 pixels wide.
\item \code{gaze_y}: Vertical pixal coordinate of fixation location (0 = top of display; 768 = bottom of display)
\item \code{fix_duration}: Duration of fixation in msec
\item \code{letter_n}: Fixated letter position in sentence (including spaces), even if the fixation was to the left of the left margin or to the right of the last letter. Position 0 is directly to the left of the first letter.
\item \code{word_n}: Position of the word in the sentence
\item \code{word}: Presented word
\item \code{blink}: Indicates whether a blink was detected directly 'before' or 'after' the current fixation (or 'both' before and after).
}
}
\usage{
data_frank2013_et_fix
}
\description{
Fixations positions of the eye-tracking  data of \insertCite{Frank2013;textual}{pangoling}.
}
\details{
As in the original dataset: "Within each sentence, fixations are listed in the order in which they occurred. If any fixation during sentence presentation was registered as being outside the display bounds, this was regarded as a tracking error and no data is presented for that sentence. The letter and word position are NaN if gaze_y is outside the range (300,475) because that was considered too far above or below the presented sentence."
}
\seealso{
\link{data_frank2013_stimuli} for the enriched stimuli, \link{data_frank2013_spr_complete} for the associated self-paced reading data, \link{data_frank2013_spr} for a cleaned version of that dataaset,  and \link{data_frank2013_et_rt} for a preprocessed this dataset.
}
\keyword{datasets}
