% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_frank2013_et_rt}
\alias{data_frank2013_et_rt}
\title{Eye-tracking data of Frank, Fernandez Monsalve, Thompson et al. (2013)}
\format{
A data frame with 81109 rows and 11 columns:
\itemize{
\item \code{subj}: Subject
\item \code{sent_id}: Sentence unique id.
\item \code{sent_n}: Position of sentence in presentation sequence
\item \code{answer_time}: Time in msec between question presentation and response, or NaN if there was no question
\item \code{acc_comprehension}: Whether the answer is correct or not
\item \code{word_n}: Position of the word in the sentence
\item \code{word}: Presented word
\item \code{RTfirstfix}: First fixation time on current word (or 0 if word not fixated)
\item \code{RTfirstpass}: First-pass reading time (total fixation time on current word before first fixation on any other word)
\item \code{RTrightbound}: Right-bounded reading time (total fixation time on current word before first fixation on any word to the right)
\item \code{RTgopast}: Go-past reading time (total fixation time from first fixation on current word up to first fixation on any word to the right)
}
}
\usage{
data_frank2013_et_rt
}
\description{
Eye-tracking reading data of Frank, Fernandez Monsalve, Thompson et al. (2013)
}
\references{
\link{1}
S. L. Frank, I. Fernandez Monsalve, R. L. Thompson, et al.
``Reading Time Data for Evaluating Broad-Coverage Models of English Sentence Processing''.
In: \emph{Behavior Research Methods} 45.4 (dec. 01, 2013), pp. 1182--1190.
ISSN: 1554-3528.
DOI: 10.3758/s13428-012-0313-y.
}
\seealso{
\link{data_frank2013_stimuli} for the enriched stimuli,  \link{data_frank2013_spr_complete} for the complete version of this dataset, \link{data_frank2013_spr} for a "cleaned" version of the self-paced reading data, and \link{data_frank2013_et_fix} for the raw fixation of this eye-tracking dataset.

Other datasets: 
\code{\link{data_frank2013_spr_complete}},
\code{\link{data_frank2013_spr}},
\code{\link{data_frank2013_stimuli}},
\code{\link{data_provo_cloze}}
}
\concept{datasets}
\keyword{datasets}
