% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_frank2013_spr_complete}
\alias{data_frank2013_spr_complete}
\title{SPR  of Frank et al ...}
\format{
A data frame with  rows and  variables:
\itemize{
\item \code{subj}: Subject
\item \code{sent_id}: Sentence unique id.
\item \code{sent_n}: Position of sentence in presentation sequence
\item \code{answer_time}: Time in msec between question presentation and response, or NaN if there was no question
\item \code{word_n}: Position of the word in the sentence
\item \code{word}: Presented word
\item \code{RT}: Time in msec between word presentation and key press
\item \code{acc_comprehension}: Correctness of response to comprehension question
\item \code{age}: Subject's age in years
\item \code{age_en}: Age at which subject began learning English (0 for native speakers)
\item \code{sex}: Subject's sex (f/m)
\item \code{hand}: Subject's handedness (r/l)
\item \code{correct_perc}: Fraction of correct responses to comprehension questions.
\item \code{typo}: Whether there is a typo in the word.
}

REFERENCES
\link{1} Frank, S.L., Fernandez Monsalve, I., Thompson, R.L, & Vigliocco, G. (2018). Reading-time data for evaluating broad-coverage models of English sentence processing. Behavior Research Methods, 45, 1182-1190.
\link{2} Fernandez Monsalve, I., Frank, S.L., & Vigliocco, G. (2012). Lexical surprisal as a general predictor of reading time. Proceedings of the 13th Conference of the European Chapter of the Association for Computational Linguistics (pp. 398-408). Avignon, France: Association for Computational Linguistics.
\link{3} Frank, S.L. (in press). Uncertainty reduction as a measure of cognitive load in sentence comprehension. Topics in Cognitive Science.
\link{4} Frank, S.L. & Thompson, R.L. (2012). Early effects of word surprisal on pupil size during reading. In: N. Miyake, D. Peebles, & R.P. Cooper (Eds.), Proceedings of the 34th Annual Conference of the Cognitive Science Society (pp. 1554-1559). Austin, TX: Cognitive Science Society.
Frank, S.L. (2013). Uncertainty reduction as a measure of cognitive load in sentence comprehension. Topics in Cognitive Science.
}
\usage{
data_frank2013_spr_complete
}
\description{
A dataset containing the .. reference
NOTE -- After collecting the self-paced-reading data, three typo's were found in the sentence stimuli:
sentence 43  : "Scott" was "Sott"
sentence 269 : "at" was "that"
sentence 337 : "Margaret" was "Margeret" These errors were fixed in the eye-tracking study.
}
\keyword{datasets}
