% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_frank2013_stimuli}
\alias{data_frank2013_stimuli}
\title{Enriched  stimuli of Frank, Fernandez Monsalve, Thompson, and Vigliocco (2013) data}
\format{
A data frame with  rows and  variables:
\itemize{
\item \code{sent_id}: The sentence unique id.
\item \code{sentence}: The sentence, as a single string
\item \code{question}: Comprehension question (or hyphen if no question)
\item \code{answer}: Correct answer to comprehension question (or hyphen if no question)
\item \code{pos}: String of part-of-speech tags (Penn Treebank style)
\item \code{word_n}: Position of the word in the sentence
\item \code{word}: Presented word
\item \code{surprisal_1}-\code{surprisal_10}: Surprisal estimate at ten points (x=1 to x=10) during network training.
\item \code{deltaH1_1}-\code{deltaH4_10}: Entropy-reduction estimate at ten points during network training, with lookahead distance n.
}
}
\source{
\url{https://osf.io/sjefs//}
}
\usage{
data_frank2013_stimuli
}
\description{
A dataset containing the stimuli used in Frank, Fernandez Monsalve, Thompson et al. (2013). It is enriched with the estimates of surprisal and entropy reduction of Frank (2013).
}
\details{
As in the original dataset part-of-speech tags were generated by Tsuruoka & Tsujii's (2005) automatic tagger, and were corrected by hand (in accordance with the Penn Treebank part-of-speech  guidelines; Santorini, 1991); Punctuation marks receive their own tag, and clitics are split into two tags (e.g., "doesn't" is tagged "VBZ RB").
}
\references{
\link{1}
S. L. Frank.
``Uncertainty Reduction as a Measure of Cognitive Load in Sentence Comprehension''.
In: \emph{Topics in Cognitive Science} 5.3 (2013), pp. 475--494.
ISSN: 1756-8765.
DOI: 10.1111/tops.12025.

\link{2}
S. L. Frank, I. Fernandez Monsalve, R. L. Thompson, et al.
``Reading Time Data for Evaluating Broad-Coverage Models of English Sentence Processing''.
In: \emph{Behavior Research Methods} 45.4 (dec. 01, 2013), pp. 1182--1190.
ISSN: 1554-3528.
DOI: 10.3758/s13428-012-0313-y.
}
\seealso{
\link{data_frank2013_spr_complete} for the associated self-paced reading data, \link{data_frank2013_spr} for a cleaned version of that dataaset,  \link{data_frank2013_et_fix} for the eye-tracking version of this dataset with the fixation information, and \link{data_frank2013_et_rt} for a preprocessed eye-tracking dataset.

Other datasets: 
\code{\link{data_frank2013_et_rt}},
\code{\link{data_frank2013_spr_complete}},
\code{\link{data_frank2013_spr}},
\code{\link{data_provo_cloze}}
}
\concept{datasets}
\keyword{datasets}
