% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_provo_cloze}
\alias{data_provo_cloze}
\title{Cloze probabilities from the Provo Corpus.}
\format{
A data frame with 2730 rows and 16 variables:
\itemize{
\item word_unique_id explain
\item text_id explain
\item word_number explain
\item word explain
\item word_cleaned explain
\item word_length explain
\item total_response_count explain
\item cp explain
\item modal_response explain
\item pos_claws explain
\item word_content_or_function explain
\item word_pos explain
\item pos_cp explain
\item lsa_context_score explain
\item lsa_response_match_score explain
\item text explain
}
}
\source{
\url{https://osf.io/sjefs//}
}
\usage{
data_provo_cloze
}
\description{
A dataset containing the .. reference
}
\keyword{datasets}
