% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{df_sent}
\alias{df_sent}
\title{Example dataset: Two word-by-word sentences}
\format{
A data frame with 15 rows and 2 columns:
\describe{
\item{sent_n}{(integer) Sentence number, indicating which sentence each
word belongs to.}
\item{word}{(character) Words from the sentences.}
}
}
\usage{
df_sent
}
\description{
This dataset contains two example sentences, split
word-by-word. It is structured to demonstrate the use of the \code{pangoling}
package for processing text data.
}
\examples{
# Load the dataset
data("df_sent")
df_sent
}
\seealso{
Other datasets: 
\code{\link{df_jaeger14}}
}
\concept{datasets}
\keyword{datasets}
