% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_causal.R
\name{get_causal_log_prob_mat}
\alias{get_causal_log_prob_mat}
\title{Get a matrix with log probability of each word phrase of a vector given its previous context using a transformer model from huggingface.com}
\usage{
get_causal_log_prob_mat(
  x,
  .by = rep(1, length(x)),
  model = "gpt2",
  add_bos_token = NULL,
  stride = 1,
  config_model = NULL,
  config_tokenizer = NULL
)
}
\arguments{
\item{x}{Vector of words, phrases or texts.}

\item{.by}{Vector that indicates how the text should be split.}

\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}

\item{add_bos_token}{Whether to include beginning of text special tokens. See details}

\item{config_model}{list with other arguments that control how the model from hugging face is accessed.}

\item{config_tokenizer}{list with arguments that control how the model from hugging face is accessed.}
}
\value{
matrix
}
\description{
Get a matrix with log probability of each word phrase of a vector given its previous context using a transformer model from huggingface.com
}
