% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_causal.R
\name{get_causal_log_prob_mat}
\alias{get_causal_log_prob_mat}
\title{Title}
\usage{
get_causal_log_prob_mat(x, by = rep(1, length(x)), model = "gpt2", eot = 0)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Title
}
