% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_masked.R
\name{get_masked_log_prob_mat}
\alias{get_masked_log_prob_mat}
\title{Title}
\usage{
get_masked_log_prob_mat(
  x,
  by = rep(1, length(x)),
  model = "distilbert-base-uncased",
  n = 0
)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Title
}
