% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_utils.R
\name{get_tokens}
\alias{get_tokens}
\title{Tokenize the input}
\usage{
get_tokens(x, model = "gpt2")
}
\arguments{
\item{x}{input}

\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}
}
\value{

}
\description{
Tokenize the input
}
