% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_utils.R
\name{get_tr_vocab}
\alias{get_tr_vocab}
\title{Returns the vocabulary of a model}
\usage{
get_tr_vocab(model = "gpt2", add_special_tokens = NULL, config = NULL)
}
\arguments{
\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}

\item{add_special_tokens}{Whether to include beginning of text special tokens. By default  acts as the \href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoTokenizer}{AutoTokenizer}.}
}
\value{
A vector with the vocabulary of a model
}
\description{
Returns the vocabulary of a model
}
