% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{get_word_fpmw}
\alias{get_word_fpmw}
\title{Get the word frequency per million words of a string}
\usage{
get_word_fpmw(
  x,
  source = "SUBTLEX_US",
  ignore_case = TRUE,
  remove_punc = "[[:punct:][:blank:]]"
)
}
\arguments{
\item{remove_punc}{}
}
\value{

}
\description{
Get the word frequency per million words of a string
}
\examples{
 get_word_fpmw(x = c("car", "cat", "bubble"))

}
