% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtlex.R
\name{get_word_freq_tbl}
\alias{get_word_freq_tbl}
\alias{get_all_word_freq_tbl}
\title{Get a table with word frequencies.}
\usage{
get_word_freq_tbl(
  x,
  source = NULL,
  ignore_case = TRUE,
  remove_punc = "[[:punct:][:blank:]]",
  regex = FALSE
)

get_all_word_freq_tbl(source = NULL)
}
\arguments{
\item{x}{a vector of words.}

\item{source}{source of the frequency values. By default uses the locale of the system. Either a source name from \code{freq_sources()} or a list with the approriate constraints e.g., \code{list(language = "English")}.}

\item{ignore_case}{ignores case.}

\item{remove_punc}{removes punctuation.}

\item{regex}{allows for regex.}
}
\value{
a data frame.
}
\description{
Produces a table with word frequencies from subtlex EXPLAIN ABOUT CITATIONS.
}
\details{
Frequencies extracted from subtlex return (a subset) of the following columns:
\itemize{
\item word. This starts with a capital when the word more often starts with an uppercase letter than with a lowercase letter.
\item freq_count. This is the number of times the word appears in the corpus (i.e., on the total of 51 million words).
\item cd_count. This is the number of films in which the word appears (i.e., it has a maximum value of 8,388).
\item freq_count_lc. This is the number of times the word appears in the corpus starting with a lowercase letter. This allows users to further match their stimuli.
\item cd_count_lc. This is the number of films in which the word appears starting with a lowercase letter.
\item freq_per_million. This is the word frequency per million words. It is the measure you would preferably use in your manuscripts, because it is a standard measure of word frequency independent of the corpus size. It is given with two digits precision, in order not to lose precision of the frequency counts.
\item log10freq_count. This value is based on log10(freq_count+1) and has four digit precision.
\item cd_percent. indicates in how many percent of the films the word appears. This value has two-digit precision in order not to lose information.
*log10_cd. This value is based on log10(CDcount+1) and has four digit precision. It is the best value to use if you want to match words on word frequency.
\item dom_pos. The dominant (most frequent) PoS of each entry
\item freq_dom_pos. The frequency of the dominant PoS
\item dom_pos_perc. The relative frequency of the dominant PoS
\item all_pos. All PoS observed for the entry
\item freq_all_pos.The frequencies of each PoS
}
}
\seealso{
Other frequency: 
\code{\link{freq_sources}()}
}
\concept{frequency}
