% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installed_py_pangoling}
\alias{installed_py_pangoling}
\title{Check if the required Python dependencies for \code{pangoling} are installed}
\usage{
installed_py_pangoling()
}
\value{
A logical value: \code{TRUE} if both \code{transformers} and \code{torch} are
installed and accessible, otherwise \code{FALSE}.
}
\description{
This function verifies whether the necessary Python modules (\code{transformers}
and \code{torch}) are available in the current Python environment.
}
\examples{
\dontrun{
if (installed_py_pangoling()) {
 message("Python dependencies are installed.")
} else {
 warning("Python dependencies are missing. Please install `torch` and `transformers`.")
}
}
}
