% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{masked_lp-defunct}
\alias{masked_lp-defunct}
\title{Get the log probability of a target word (or phrase) given a left and right context}
\description{
This function is defunct. Use \code{masked_targets_pred()} instead.
}
\seealso{
\code{\link{pangoling-defunct}}
}
\keyword{internal}
