% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{masked_tokens_tbl-defunct}
\alias{masked_tokens_tbl-defunct}
\title{Get the possible tokens and their log probabilities for each mask in a sentence}
\value{
This function is defunct and returns an error.
}
\description{
This function is defunct. Use \code{masked_tokens_pred_tbl()} instead.
}
\seealso{
\code{\link{pangoling-defunct}}
}
\keyword{internal}
