% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_masked.R
\name{masked_tokens_tbl}
\alias{masked_tokens_tbl}
\title{Get the possible tokens and their log probabilities for each mask in a sentence.}
\usage{
masked_tokens_tbl(
  masked_sentences,
  model = getOption("pangoling.masked.default"),
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL
)
}
\arguments{
\item{model}{Name of a pretrained model stored locally on the (huggingface.co).}

\item{add_special_tokens}{Whether to include beginning of text special tokens. By default  acts as the \href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoTokenizer}{AutoTokenizer}.}

\item{config_model}{List with other arguments that control how the model from hugging face is accessed.}

\item{config_tokenizer}{List with other arguments that control how the tokenizer from hugging face is accessed.}

\item{masked_sentence}{Context}
}
\value{
A table with
}
\description{
For each mask in a sentence, get the possible tokens and their log probabilities using a masked transformer
}
\details{
For more about masked models, see \href{https://huggingface.co/course/chapter?/?}{chapter ? of hugging face documentation}.

If not specified, the causal model that will be used is the one set in specified in the global option \code{pangoling.masked.default}, this can be accessed via \code{getOption("pangoling.masked.default")} (by default "bert-base-uncased"). To change the default option use \code{options(pangoling.masked.default = "newmaskedmodel")}.

A list of possible causal masked can be found in \href{https://huggingface.co/}{hugging face website}.

Using the  \code{config_model} and \code{config_tokenizer} arguments, it's possible to control how the model and tokenizer from hugging face is accessed, see the python method \href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoProcessor.from_pretrained}{\code{from_pretrained}} for details. In case of errors check the status of \url{https://status.huggingface.co/}
}
\section{More examples}{

See the  \href{https://bruno.nicenboim.me/pangoling/articles/intro-bert.html}{online article} in pangoling website for more examples.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
masked_tokens_tbl("The [MASK] doesn't fall far from the tree.",
                   model = "bert-base-uncased")
\dontshow{\}) # examplesIf}
}
\seealso{
Other masked model functions: 
\code{\link{causal_config}()},
\code{\link{masked_preload}()}
}
\concept{masked model functions}
