% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_utils.R
\name{ntokens}
\alias{ntokens}
\title{The number of tokens in a string or vector of strings}
\usage{
ntokens(
  x,
  model = getOption("pangoling.causal.default"),
  add_special_tokens = NULL,
  config_tokenizer = NULL
)
}
\arguments{
\item{x}{character input}
}
\value{
The number of tokens in a string or vector of words.
}
\description{
The number of tokens in a string or vector of strings
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ntokens(x = c("The apple doesn't fall far from the tree."), model = "gpt2")
\dontshow{\}) # examplesIf}
}
\seealso{
Other token-related functions: 
\code{\link{tokenize_lst}()},
\code{\link{transformer_vocab}()}
}
\concept{token-related functions}
