% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_utils.R
\name{ntokens}
\alias{ntokens}
\title{The number of tokens in a string or vector of strings}
\usage{
ntokens(x, model = "gpt2", add_bos_token = NULL, config = NULL)
}
\arguments{
\item{x}{character input}

\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}

\item{add_bos_token}{Whether to include beginning of text special tokens. See details}
}
\value{
The number of tokens in a string or vector of words.
}
\description{
The number of tokens in a string or vector of strings
}
