% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangoling-package.R
\docType{package}
\name{pangoling-package}
\alias{pangoling}
\alias{pangoling-package}
\title{pangoling: Access to Large Language Model Predictions}
\description{
Access to word predictability using large language (transformer) models.
}
\details{
These options are used to control various aspects of the pangoling package.
Users can customize these options via the \code{options()} function by specifying \verb{pangoling.<option>} names.
\itemize{
\item \code{pangoling.debug}: Logical; when \code{TRUE}, enables debugging mode. Default is \code{FALSE}.
\item \code{pangoling.verbose}: Integer; controls the verbosity level (e.g., 0 = silent, 1 = minimal, 2 = detailed). Default is \code{2}.
\item \code{pangoling.log.p}: Logical; if \code{TRUE}, probabilities are logged. Default is \code{TRUE}.If TRUE (default),  x are assumed to be log-transformed probabilities with base e, if FALSE x are assumed to be raw probabilities, alternatively log.p can be the base of other logarithmic transformations.
\item \code{pangoling.cache}: A cache object created with \code{cachem::cache_mem}, allowing you to specify the maximum size (in bytes) for cached objects. Default is \code{1024 * 1024^2} bytes (1 MB).
\item \code{pangoling.causal.default}: Character string; specifies the default model for causal language processing. Default is \code{"gpt2"}.
\item \code{pangoling.masked.default}: Character string; specifies the default model for masked language processing. Default is \code{"bert-base-uncased"}.
}

Use \verb{options(pangoling.<option> = <value>)} to set these options in your session.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bruno.nicenboim.me/pangoling}
  \item \url{https://github.com/bnicenboim/pangoling}
  \item Report bugs at \url{https://github.com/bnicenboim/pangoling/issues}
}

}
\author{
\strong{Maintainer}: Bruno Nicenboim \email{bruno.nicenboim@gmail.com} (\href{https://orcid.org/0000-0002-5176-3943}{ORCID})

Other contributors:
\itemize{
  \item Chris Emmerly [contributor]
  \item Giovanni Cassani [contributor]
  \item Lisa Levinson [reviewer]
  \item Utku Turk [reviewer]
}

}
\keyword{internal}
