% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangoling-package.R
\docType{package}
\name{pangoling-package}
\alias{pangoling}
\alias{pangoling-package}
\title{pangoling: Access to Language Model Predictions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access to word predictability using large language (transformer) models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bruno.nicenboim.me/pangoling}
  \item \url{https://github.com/bnicenboim/pangoling}
  \item Report bugs at \url{https://github.com/bnicenboim/pangoling/issues}
}

}
\author{
\strong{Maintainer}: Bruno Nicenboim \email{bruno.nicenboim@gmail.com} (\href{https://orcid.org/0000-0002-5176-3943}{ORCID})

Other contributors:
\itemize{
  \item Chris Emmerly [contributor]
  \item Giovanni Cassani [contributor]
}

}
\keyword{internal}
