% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_causal.R
\name{preload_causal}
\alias{preload_causal}
\title{Preloads a causal language model}
\usage{
preload_causal(
  model = "gpt2",
  add_bos_token = NULL,
  config_model = NULL,
  config_tokenizer = NULL
)
}
\arguments{
\item{model}{Name of a pretrained model stored on the huggingface.co. (Maybe a path to a  model (.pt or .bin file) stored locally will work.)}

\item{add_bos_token}{Whether to include beginning of text special tokens. By default  acts as the \href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoTokenizer}{AutoTokenizer}.}

\item{config_model}{List with other arguments that control how the model from hugging face is accessed.}

\item{config_tokenizer}{List with arguments that control how the tokenizer from hugging face is accessed.}
}
\value{
Nothing.
}
\description{
Preloads a causal language model to speed up next runs.
}
