% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_utils.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize the input}
\usage{
tokenize(x, model = "gpt2", add_special_tokens = NULL, config_tokenizer = NULL)
}
\arguments{
\item{x}{Strings or token ids.}

\item{model}{Name of a pretrained model stored locally on the (huggingface.co).}

\item{add_special_tokens}{Whether to include beginning of text special tokens. By default  acts as the \href{https://huggingface.co/docs/transformers/v4.25.1/en/model_doc/auto#transformers.AutoTokenizer}{AutoTokenizer}.}

\item{config_tokenizer}{List with other arguments that control how the tokenizer from hugging face is accessed.}

\item{config}{List with arguments that control how the tokenizer from hugging face is accessed.}
}
\description{
Tokenize the input
}
