% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_utils.R
\name{tokenize_lst}
\alias{tokenize_lst}
\title{Tokenize an input}
\usage{
tokenize_lst(
  x,
  model = getOption("pangoling.causal.default"),
  add_special_tokens = NULL,
  config_tokenizer = NULL
)
}
\arguments{
\item{x}{Strings or token ids.}
}
\value{
A list with tokens
}
\description{
Tokenize a string or token ids.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tokenize_lst(x = c("The apple doesn't fall far from the tree."), model = "gpt2")
\dontshow{\}) # examplesIf}
}
\seealso{
Other token-related functions: 
\code{\link{ntokens}()},
\code{\link{transformer_vocab}()}
}
\concept{token-related functions}
